behave-ui-hotkeys now uses the UMD format for modules
- All src moved to a single file
- added a bower.json file for browser users

Conflicts:
	.gitignore
	src/behave-ui-hotkeys.js
